+++ CHANGELOG.md
@@ -6,6 +6,24 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: Add missing exports to ESM exports.
 
+Breaking changes:
+
+- Delete the following deprecated SCSS mixins and styles:
+  - `k-Dropdown`
+  - `k-PlatformSwitch`
+  - `k-UserMenu`
+- Move the following components:
+  - `Dropdown` -> `components/navigation/header-menu/components`
+  - `DropdownButton` -> `components/navigation/header-menu/components`
+- Remove default export for `Dropdown` and `DropdownButton`.
+- Remove unused props for `Dropdown`.
+
+- Feature: Add the following Logo components:
+  - `MaisonDeCrowdfundingLogo`
+  - `MicrodonLogo`
+  - `GoodeedLogo`
+  - `KkbbCoLogo`
+  
 ## [3.0.0-beta.18] - 2021-03-18
 
 Fix: